It works don't touch anything, don't ask me to explain it

+++ ios/ARView.swift
@@ -12,7 +12,7 @@ import SceneKit
 class ARView : ARSCNView, ARSCNViewDelegate {
     
     var objects = [SCNNode]()
-    var objectHashTable = [Int:SCNNode]()
+    var objectHashTable = [Int64:SCNNode]()
     var objectToPlace: String?
     
     var selectedNode: SCNNode?
